[review Fix] Modify the variables and abstract the request queue into a separate file (#22643)
